Rename getIsBlockActive() to isBlockActive() + move more funcs to helpers
