+++ src/blocks/homepage-articles/edit.js
@@ -5,13 +5,12 @@
  */
 import QueryControls from '../../components/query-controls';
 import { STORE_NAMESPACE } from './store';
-import { getIsBlogPrivate } from './helpers';
+import { isBlogPrivate, isSpecificPostModeActive, queryCriteriaFromAttributes } from './helpers';
 
 /**
  * External dependencies
  */
 import classNames from 'classnames';
-import { isUndefined, pickBy } from 'lodash';
 
 /**
  * WordPress dependencies
@@ -332,7 +331,7 @@ class Edit extends Component {
 							required
 						/>
 					) }
-					{ ! specificMode && ! getIsBlogPrivate() && (
+					{ ! specificMode && ! isBlogPrivate() && (
 						/*
 						 * Hide the "More" button option on private sites.
 						 *
@@ -642,7 +641,7 @@ class Edit extends Component {
 					</div>
 				</div>
 
-				{ ! specificMode && latestPosts && moreButton && ! getIsBlogPrivate() && (
+				{ ! specificMode && latestPosts && moreButton && ! isBlogPrivate() && (
 					/*
 					 * The "More" button option is hidden for private sites, so we should
 					 * also hide the button in case it was previously enabled.
@@ -671,30 +670,6 @@ class Edit extends Component {
 		);
 	}
 }
-
-const isSpecificPostModeActive = ( { specificMode, specificPosts } ) =>
-	specificMode && specificPosts && specificPosts.length;
-
-const queryCriteriaFromAttributes = attributes => {
-	const { postsToShow, authors, categories, tags, specificPosts, tagExclusions } = attributes;
-	const criteria = pickBy(
-		isSpecificPostModeActive( attributes )
-			? {
-					include: specificPosts,
-					orderby: 'include',
-					per_page: specificPosts.length,
-			  }
-			: {
-					per_page: postsToShow,
-					categories,
-					author: authors,
-					tags,
-					tags_exclude: tagExclusions,
-			  },
-		value => ! isUndefined( value )
-	);
-	return criteria;
-};
 
 export default compose( [
 	withColors( { textColor: 'color' } ),
